+++ src/lib.rs
@@ -47,6 +47,7 @@ assert_eq!(path.with_write_opt(&opt).to_string(), path_str);
 | [\<color\>]               | Color         | Stack   |                     |
 | [\<number\>]              | f64           | Stack   |                     |
 | [\<length\>]              | Length        | Stack   |                     |
+| [\<angle\>]               | Angle         | Stack   |                     |
 | [\<viewBox\>]             | ViewBox       | Stack   |                     |
 | [\<path\>]                | Path          | Heap    | PathParser          |
 | [\<list-of-numbers\>]     | NumberList    | Heap    | NumberListParser    |
@@ -58,6 +59,7 @@ assert_eq!(path.with_write_opt(&opt).to_string(), path_str);
 [\<color\>]: https://www.w3.org/TR/SVG11/types.html#DataTypeColor
 [\<number\>]: https://www.w3.org/TR/SVG11/types.html#DataTypeNumber
 [\<length\>]: https://www.w3.org/TR/SVG11/types.html#DataTypeLength
+[\<angle\>]: https://www.w3.org/TR/SVG11/types.html#DataTypeAngle
 [\<viewBox\>]: https://www.w3.org/TR/SVG11/coords.html#ViewBoxAttribute
 [\<path\>]: https://www.w3.org/TR/SVG11/paths.html#PathData
 [\<list-of-numbers\>]: https://www.w3.org/TR/SVG11/types.html#DataTypeList
@@ -99,7 +101,7 @@ assert_eq!(path.with_write_opt(&opt).to_string(), path_str);
 None.
 */
 
-#![doc(html_root_url = "https://docs.rs/svgtypes/0.4.4")]
+#![doc(html_root_url = "https://docs.rs/svgtypes/0.5.0")]
 
 #![forbid(unsafe_code)]
 #![deny(missing_docs)]