+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 - `PathBuilder`. Use `Path::push_*` instead.
 - `Style` parser. Use an external CSS parser instead, like `simplecss`.
+- `ElementId` and `AttributeId`.
 
 ## [0.4.4] - 2019-06-11
 - Update `float-cmp`.