+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.5.0] - 2019-08-12
 ### Added
 - Implement `Default` for `Length`, `LengthList`, `NumberList`, `Points` and `Path`.
 
@@ -80,7 +82,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - `Paint::from_span` poor performance.
 
-[Unreleased]: https://github.com/RazrFalcon/svgtypes/compare/v0.4.4...HEAD
+[Unreleased]: https://github.com/RazrFalcon/svgtypes/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/RazrFalcon/svgtypes/compare/v0.4.4...v0.5.0
 [0.4.4]: https://github.com/RazrFalcon/svgtypes/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/RazrFalcon/svgtypes/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/RazrFalcon/svgtypes/compare/v0.4.1...v0.4.2